fix all bug in register controller

+++ src/utils/cloudinary.js
@@ -15,12 +15,19 @@ const uploadOnCloudinary = async (localFilePath) => {
       resource_type: 'auto',
     });
     // file has been uploaded successfully
-    console.log('file is uploaded successfully', response.url);
+    // console.log('file is uploaded successfully', response.url);
+    fs.unlinkSync(localFilePath);
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
+    console.error('Cloudinary Error:');
+    console.error(error);
+
+    if (localFilePath && fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+
     return null;
   }
 };
 
-export {uploadOnCloudinary}
+export { uploadOnCloudinary };
